added categories component
