+++ src/Product.js
@@ -3,7 +3,6 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 
 const Product = ({ shoe }) => {
-  console.log(shoe);
   const { img, title, discount, price } = shoe;
   const newPrice = price - price * discount;
